+++ okta_jwt_verifier/__init__.py
@@ -2,6 +2,7 @@
 
 Allow to verify JWT locally
 """
-__version__ = '0.1.0'
+__version__ = '0.2.0'
 
-from .jwt_verifier import JWTVerifier # noqa
+from .jwt_verifier import JWTVerifier, AccessTokenVerifier, IDTokenVerifier # noqa
+from .jwt_utils import JWTUtils # noqa