+++ okta_jwt_verifier/jwt_verifier.py
@@ -1,27 +1,28 @@
-import json
+import warnings
 
 from urllib.parse import urljoin
-from jose import jwt, jws
 
 from . import __version__ as version
 from .config_validator import ConfigValidator
 from .constants import MAX_RETRIES, MAX_REQUESTS, REQUEST_TIMEOUT, LEEWAY
 from .exceptions import JWKException, JWTValidationException
+from .jwt_utils import JWTUtils
 from .request_executor import RequestExecutor
 
 
 class JWTVerifier():
 
     def __init__(self,
-                 issuer,
-                 client_id,
+                 issuer=None,
+                 client_id='client_id_stub',
                  audience='api://default',
                  request_executor=RequestExecutor,
                  max_retries=MAX_RETRIES,
                  request_timeout=REQUEST_TIMEOUT,
                  max_requests=MAX_REQUESTS,
                  leeway=LEEWAY,
-                 cache_jwks=True):
+                 cache_jwks=True,
+                 proxy=None):
         """
         Args:
             issuer: string, full URI of the token issuer, required
@@ -34,6 +35,10 @@ class JWTVerifier():
             leeway: int, amount of time to expand the window for token expiration (to work around clock skew)
             cache_jwks: bool, optional
         """
+        warnings.simplefilter('module')
+        warnings.warn('JWTVerifier will be deprecated soon. '
+                      'For token verification use IDTokenVerifier or AccessTokenVerifier. '
+                      'For different jwt utils use JWTUtils.', DeprecationWarning)
         # validate input data before any processing
         config = {'issuer': issuer,
                   'client_id': client_id,
@@ -50,7 +55,8 @@ class JWTVerifier():
         self.audience = audience
         self.request_executor = request_executor(max_retries=max_retries,
                                                  max_requests=max_requests,
-                                                 request_timeout=request_timeout)
+                                                 request_timeout=request_timeout,
+                                                 proxy=proxy)
         self.max_retries = max_retries
         self.request_timeout = request_timeout
         self.max_requests = max_requests
@@ -63,9 +69,7 @@ class JWTVerifier():
         Return:
             tuple (headers, claims, signing_input, signature)
         """
-        headers, payload, signing_input, signature = jws._load(token)
-        claims = json.loads(payload.decode('utf-8'))
-        return (headers, claims, signing_input, signature)
+        return JWTUtils.parse_token(token)
 
     async def verify_access_token(self, token, claims_to_verify=('iss', 'aud', 'exp')):
         """Verify acess token.
@@ -156,39 +160,19 @@ class JWTVerifier():
 
     def verify_signature(self, token, okta_jwk):
         """Verify token signature using received jwk."""
-        headers, claims, signing_input, signature = self.parse_token(token)
-        jws._verify_signature(signing_input=signing_input,
-                              header=headers,
-                              signature=signature,
-                              key=okta_jwk,
-                              algorithms=['RS256'])
+        JWTUtils.verify_signature(token, okta_jwk)
 
     def verify_claims(self, claims, claims_to_verify, leeway=LEEWAY):
         """Verify claims are present and valid."""
-        # Check if required claims are present, because library "jose" doesn't raise an exception
-        for claim in claims_to_verify:
-            if claim not in claims:
-                raise JWTValidationException(f'Required claim "{claim}" is not present.')
-
-        # Overwrite defaults in python-jose library
-        options = {'verify_aud': 'aud' in claims_to_verify,
-                   'verify_iat': 'iat' in claims_to_verify,
-                   'verify_exp': 'exp' in claims_to_verify,
-                   'verify_nbf': 'nbf' in claims_to_verify,
-                   'verify_iss': 'iss' in claims_to_verify,
-                   'verify_sub': 'sub' in claims_to_verify,
-                   'verify_jti': 'jti' in claims_to_verify,
-                   'leeway': leeway}
-        # Validate claims
-        jwt._validate_claims(claims,
-                             audience=self.audience,
-                             issuer=self.issuer,
-                             options=options)
+        JWTUtils.verify_claims(claims,
+                               claims_to_verify,
+                               self.audience,
+                               self.issuer,
+                               leeway)
 
     def verify_expiration(self, token, leeway=LEEWAY):
         """Verify if token is not expired."""
-        headers, claims, signing_input, signature = self.parse_token(token)
-        self.verify_claims(claims, claims_to_verify=('exp'), leeway=LEEWAY)
+        JWTUtils.verify_expiration(token, leeway)
 
     def _get_jwk_by_kid(self, jwks, kid):
         """Loop through given jwks and find jwk which matches by kid.
@@ -255,3 +239,79 @@ class JWTVerifier():
     def _clear_requests_cache(self):
         """Clear whole cache."""
         self.request_executor.clear_cache()
+
+
+class AccessTokenVerifier():
+    def __init__(self,
+                 issuer=None,
+                 audience='api://default',
+                 request_executor=RequestExecutor,
+                 max_retries=MAX_RETRIES,
+                 request_timeout=REQUEST_TIMEOUT,
+                 max_requests=MAX_REQUESTS,
+                 leeway=LEEWAY,
+                 cache_jwks=True,
+                 proxy=None):
+        """
+        Args:
+            issuer: string, full URI of the token issuer, required
+            audience: string, expected audience, optional
+            request_executor: RequestExecutor class or its subclass, optional
+            max_retries: int, number of times to retry a failed network request, optional
+            request_timemout: int, max request timeout, optional
+            max_requests: int, max number of concurrent requests
+            leeway: int, amount of time to expand the window for token expiration (to work around clock skew)
+            cache_jwks: bool, optional
+        """
+        self._jwt_verifier = JWTVerifier(issuer,
+                                         'client_id_stub',
+                                         audience,
+                                         request_executor,
+                                         max_retries,
+                                         request_timeout,
+                                         max_requests,
+                                         leeway,
+                                         cache_jwks,
+                                         proxy)
+
+    async def verify(self, token, claims_to_verify=('iss', 'aud', 'exp')):
+        await self._jwt_verifier.verify_access_token(token, claims_to_verify)
+
+
+class IDTokenVerifier():
+    def __init__(self,
+                 issuer=None,
+                 client_id='client_id_stub',
+                 audience='api://default',
+                 request_executor=RequestExecutor,
+                 max_retries=MAX_RETRIES,
+                 request_timeout=REQUEST_TIMEOUT,
+                 max_requests=MAX_REQUESTS,
+                 leeway=LEEWAY,
+                 cache_jwks=True,
+                 proxy=None):
+        """
+        Args:
+            issuer: string, full URI of the token issuer, required
+            client_id: string, expected client_id, required
+            audience: string, expected audience, optional
+            request_executor: RequestExecutor class or its subclass, optional
+            max_retries: int, number of times to retry a failed network request, optional
+            request_timemout: int, max request timeout, optional
+            max_requests: int, max number of concurrent requests
+            leeway: int, amount of time to expand the window for token expiration (to work around clock skew)
+            cache_jwks: bool, optional
+        """
+        self._jwt_verifier = JWTVerifier(issuer,
+                                         client_id,
+                                         audience,
+                                         request_executor,
+                                         max_retries,
+                                         request_timeout,
+                                         max_requests,
+                                         leeway,
+                                         cache_jwks,
+                                         proxy)
+
+    async def verify(self, token, claims_to_verify=('iss', 'exp'), nonce=None):
+        await self._jwt_verifier.verify_id_token(token, claims_to_verify, nonce)