+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # OKTA JWT Verifier Changelog
 
+## v0.2.0
+- Add classes IDTokenVerifier and AccessTokenVerifier
+- Mark JWTVerifier class as deprecated. This class will be removed in the next major version. 
+- Add proxy support
+- Update README
+- Few codebase improvements
+
+_New features:_
+- Separate classes for verifying ID Tokens and Access Tokens
+- Add proxy support
+
 ## v0.1.0
 - Initial release